requirements.txt
torch>=2.1
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.66

